Added a method to the tag collection to be able to get all values
Also updated the README with some configuration guidance. I felt it was necessary to have the getTagValues function since it was the easiest way to get all values.
And since there isn't any documentation on how to customize even though it says it's higly customizable was irritating. Hopefully with the documentation it will be a bit more plug and play for people. Without having to go through the source code.